# supabase_openapi.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "title": "standard public schema",
    "version": "14.5"
  },
  "host": "fwrkrpmqmqxlgvyzrizq.supabase.co:443",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json",
    "application/vnd.pgrst.object+json;nulls=stripped",
    "application/vnd.pgrst.object+json",
    "text/csv"
  ],
  "produces": [
    "application/json",
    "application/vnd.pgrst.object+json;nulls=stripped",
    "application/vnd.pgrst.object+json",
    "text/csv"
  ],
  "paths": {
...
    "/therapists": {
      "get": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.therapists.id"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.full_name"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.email"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.phone"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.qualification"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.license_number"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.years_of_experience"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.languages_spoken"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.consultation_fee"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.status"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.doctor_code"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.rating"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.updated_at"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/order"
          },
          {
            "$ref": "#/parameters/range"
          },
          {
            "$ref": "#/parameters/rangeUnit"
          },
          {
            "$ref": "#/parameters/offset"
          },
          {
            "$ref": "#/parameters/limit"
          },
          {
            "$ref": "#/parameters/preferCount"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/therapists"
              },
              "type": "array"
            }
          },
          "206": {
            "description": "Partial Content"
          }
        },
        "tags": [
          "therapists"
        ]
      },
      "post": {
        "parameters": [
          {
            "$ref": "#/parameters/body.therapists"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/preferPost"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "tags": [
          "therapists"
        ]
      },
      "delete": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.therapists.id"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.full_name"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.email"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.phone"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.qualification"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.license_number"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.years_of_experience"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.languages_spoken"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.consultation_fee"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.status"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.doctor_code"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.rating"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.updated_at"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "therapists"
        ]
      },
      "patch": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.therapists.id"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.full_name"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.email"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.phone"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.qualification"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.license_number"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.years_of_experience"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.languages_spoken"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.consultation_fee"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.status"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.doctor_code"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.rating"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.therapists.updated_at"
          },
          {
            "$ref": "#/parameters/body.therapists"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "therapists"
        ]
      }
    },
...
  },
  "definitions": {
...
    "therapists": {
      "required": [
        "id",
        "full_name",
        "email",
        "phone",
        "qualification",
        "license_number",
        "years_of_experience",
        "languages_spoken",
        "consultation_fee",
        "status",
        "rating",
        "created_at",
        "updated_at"
      ],
      "properties": {
        "id": {
          "description": "Note:\nThis is a Primary Key.<pk/>",
          "format": "text",
          "type": "string"
        },
        "full_name": {
          "default": "",
          "format": "text",
          "type": "string"
        },
        "email": {
          "default": "",
          "format": "text",
          "type": "string"
        },
        "phone": {
          "default": "",
          "format": "text",
          "type": "string"
        },
        "qualification": {
          "default": "",
          "format": "text",
          "type": "string"
        },
        "license_number": {
          "default": "",
          "format": "text",
          "type": "string"
        },
        "years_of_experience": {
          "default": 0,
          "format": "integer",
          "type": "integer"
        },
        "languages_spoken": {
          "default": "",
          "format": "text",
          "type": "string"
        },
        "consultation_fee": {
          "default": 0,
          "format": "numeric",
          "type": "number"
        },
        "status": {
          "default": "pending",
          "format": "text",
          "type": "string"
        },
        "doctor_code": {
          "format": "text",
          "type": "string"
        },
        "rating": {
          "default": 0,
          "format": "numeric",
          "type": "number"
        },
        "created_at": {
          "default": "now()",
          "format": "timestamp with time zone",
          "type": "string"
        },
        "updated_at": {
          "default": "now()",
          "format": "timestamp with time zone",
          "type": "string"
        }
      },
      "type": "object"
    },
...
  },
  "parameters": {
...
    "preferReturn": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "return=representation",
        "return=minimal",
        "return=none"
      ],
      "in": "header",
      "type": "string"
    },
    "preferCount": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "count=none"
      ],
      "in": "header",
      "type": "string"
    },
    "preferPost": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "return=representation",
        "return=minimal",
        "return=none",
        "resolution=ignore-duplicates",
        "resolution=merge-duplicates"
      ],
      "in": "header",
      "type": "string"
    },
    "select": {
      "name": "select",
      "description": "Filtering Columns",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "order": {
      "name": "order",
      "description": "Ordering",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "range": {
      "name": "Range",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "header",
      "type": "string"
    },
    "rangeUnit": {
      "name": "Range-Unit",
      "description": "Limiting and Pagination",
      "required": false,
      "default": "items",
      "in": "header",
      "type": "string"
    },
    "offset": {
      "name": "offset",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "limit": {
      "name": "limit",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "body.therapists": {
      "name": "therapists",
      "description": "therapists",
      "required": false,
      "in": "body",
      "schema": {
        "$ref": "#/definitions/therapists"
      }
    },
    "rowFilter.therapists.id": {
      "name": "id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.therapists.full_name": {
      "name": "full_name",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.therapists.email": {
      "name": "email",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.therapists.phone": {
      "name": "phone",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.therapists.qualification": {
      "name": "qualification",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.therapists.license_number": {
      "name": "license_number",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.therapists.years_of_experience": {
      "name": "years_of_experience",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.therapists.languages_spoken": {
      "name": "languages_spoken",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.therapists.consultation_fee": {
      "name": "consultation_fee",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.therapists.status": {
      "name": "status",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.therapists.doctor_code": {
      "name": "doctor_code",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.therapists.rating": {
      "name": "rating",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.therapists.created_at": {
      "name": "created_at",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.therapists.updated_at": {
      "name": "updated_at",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
  },
...
}